+++ cli/tests/test_simplify_parents_command.rs
@@ -37,21 +37,6 @@ fn create_commit(test_env: &TestEnvironment, repo_path: &Path, name: &str, paren
 }
 
 #[test]
-fn test_simplify_parents_no_args() {
-    let (test_env, repo_path) = create_repo();
-
-    let stderr = test_env.jj_cmd_cli_error(&repo_path, &["simplify-parents"]);
-    insta::assert_snapshot!(stderr, @r###"
-    error: the following required arguments were not provided:
-      <--source <SOURCE>|--revisions <REVISIONS>>
-
-    Usage: jj simplify-parents <--source <SOURCE>|--revisions <REVISIONS>>
-
-    For more information, try '--help'.
-    "###);
-}
-
-#[test]
 fn test_simplify_parents_no_commits() {
     let (test_env, repo_path) = create_repo();
 
@@ -244,6 +229,77 @@ fn test_simplify_parents_multiple_redundant_parents() {
     ○  d
     ○  c
     ○  b
+    ○  a
+    ◆
+    "#);
+}
+
+#[test]
+fn test_simplify_parents_no_args() {
+    let (test_env, repo_path) = create_repo();
+
+    create_commit(&test_env, &repo_path, "a", &["root()"]);
+    create_commit(&test_env, &repo_path, "b", &["a"]);
+    create_commit(&test_env, &repo_path, "c", &["a", "b"]);
+    create_commit(&test_env, &repo_path, "d", &["c"]);
+    create_commit(&test_env, &repo_path, "e", &["d"]);
+    create_commit(&test_env, &repo_path, "f", &["d", "e"]);
+    let stdout = test_env.jj_cmd_success(&repo_path, &["log", "-r", "all()", "-T", "description"]);
+    insta::assert_snapshot!(stdout, @r#"
+    @    f
+    ├─╮
+    │ ○  e
+    ├─╯
+    ○  d
+    ○    c
+    ├─╮
+    │ ○  b
+    ├─╯
+    ○  a
+    ◆
+    "#);
+    let setup_opid = test_env.current_operation_id(&repo_path);
+
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["simplify-parents"]);
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @r#"
+    Removed 2 edges from 2 out of 6 commits.
+    Rebased 2 descendant commits
+    Working copy now at: kmkuslsw 8cc01e1b f | f
+    Parent commit      : znkkpsqq 040ae3a6 e | e
+    "#);
+
+    let stdout = test_env.jj_cmd_success(&repo_path, &["log", "-r", "all()", "-T", "description"]);
+    insta::assert_snapshot!(stdout, @r#"
+    @  f
+    ○  e
+    ○  d
+    ○  c
+    ○  b
+    ○  a
+    ◆
+    "#);
+
+    // Test with custom `revsets.simplify-parents`.
+    test_env.jj_cmd_ok(&repo_path, &["op", "restore", &setup_opid]);
+    test_env.add_config(r#"revsets.simplify-parents = "d::""#);
+    let (stdout, stderr) = test_env.jj_cmd_ok(&repo_path, &["simplify-parents"]);
+    insta::assert_snapshot!(stdout, @"");
+    insta::assert_snapshot!(stderr, @r#"
+    Removed 1 edges from 1 out of 3 commits.
+    Working copy now at: kmkuslsw 0c6b4c43 f | f
+    Parent commit      : znkkpsqq 6a679611 e | e
+    "#);
+
+    let stdout = test_env.jj_cmd_success(&repo_path, &["log", "-r", "all()", "-T", "description"]);
+    insta::assert_snapshot!(stdout, @r#"
+    @  f
+    ○  e
+    ○  d
+    ○    c
+    ├─╮
+    │ ○  b
+    ├─╯
     ○  a
     ◆
     "#);