+++ CHANGELOG.md
@@ -71,6 +71,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   "diff3" conflict style, meaning it is more likely to work with external tools,
   but it doesn't support conflicts with more than 2 sides.
 
+* `jj simplify-parents` now supports configuring the default revset when no
+   `--source` or `--revisions` arguments are provided with the
+   `revsets.simplify-parents` config.
+
 ### Fixed bugs
 
 * `jj config unset <TABLE-NAME>` no longer removes a table (such as `[ui]`.)